refactor(protest): 단일책임원칙 준수를 위한 ProtestFacade 도입
