+++ src/main/java/com/on/eye/api/protest/controller/ProtestController.java
@@ -14,8 +14,10 @@ import org.springframework.web.bind.annotation.*;
 import com.on.eye.api.protest.dto.Coordinate;
 import com.on.eye.api.protest.dto.ProtestCreateRequest;
 import com.on.eye.api.protest.dto.ProtestResponse;
+import com.on.eye.api.protest.service.ProtestFacade;
 import com.on.eye.api.protest.service.ProtestService;
 import com.on.eye.api.protest_verification.dto.ProtestVerificationResponse;
+import com.on.eye.api.protest_verification.service.ProtestVerificationService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -25,19 +27,21 @@ import lombok.RequiredArgsConstructor;
 @Validated
 public class ProtestController {
     private final ProtestService protestService;
+    private final ProtestFacade protestFacade;
+    private final ProtestVerificationService protestVerificationService;
 
     @PostMapping("{id}/participate/verify")
     public ResponseEntity<Boolean> participateVerify(
             @PathVariable Long id, @Valid @RequestBody Coordinate request) {
-        // 냅다 ok가 아니라, 성공이면 ok, 실패면 success = false. 거리가 멀다 이런 식의 메세지가 나갔으면 함. 실패 사유가 나갔으면.
-        return ResponseEntity.ok(protestService.participateVerify(id, request));
+
+        return ResponseEntity.ok(protestFacade.participateVerify(id, request));
     }
 
     @PostMapping
     public ResponseEntity<List<Long>> createProtest(
             @Valid @RequestBody List<ProtestCreateRequest> protestCreateRequest) {
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(protestService.createProtest(protestCreateRequest));
+                .body(protestFacade.createProtest(protestCreateRequest));
     }
 
     @GetMapping("/{id}")
@@ -56,7 +60,7 @@ public class ProtestController {
                     LocalDate date) {
         if (date == null) date = LocalDate.now();
         List<ProtestVerificationResponse> response =
-                protestService.getProtestVerifications(protestId, date);
+                protestVerificationService.getTodayProtestVerifications(protestId, date);
         return ResponseEntity.ok(response);
     }
